% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{.get_builtin_models}
\alias{.get_builtin_models}
\title{Return built-in epidemiological models}
\usage{
.get_builtin_models()
}
\description{
Return built-in epidemiological models
}
\value{
A named list of built-in objects of class \code{"epi_model"}.
List names are model identifiers and values are the corresponding model definitions.
}
\keyword{internal}
