% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry_api.R
\name{list_models}
\alias{list_models}
\title{List registered epidemiological models}
\usage{
list_models()
}
\value{
A character vector containing the names of the models currently
stored in the internal registry. Each element is a model identifier that can
be passed to \code{get_model()} to retrieve the corresponding
\code{epi_model} object.
}
\description{
Returns the names of all epidemiological models currently stored
in the internal registry.
}
\details{
The registry contains:
\itemize{
\item Built-in models registered at package load time.
\item User-defined models registered via \code{\link{register_epi_model}}.
}

The returned value is a character vector containing model names.
}
\examples{
list_models()

}
\seealso{
\code{\link{register_epi_model}},
\code{\link{get_model}}
}
