% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODEL_SIR.R
\docType{data}
\name{SIR_MODEL}
\alias{SIR_MODEL}
\title{SIR epidemic model}
\format{
An object of class \code{"epi_model"}.
}
\usage{
SIR_MODEL
}
\value{
An object of class \code{"epi_model"} representing the predefined
deterministic SIR (Susceptible-Infectious-Recovered) compartmental epidemic
model. The returned object contains the model right-hand side, declared
state and derived variables, parameter names, and default initial
conditions needed by utilities such as \code{\link{simulate_epi}}.
}
\description{
An \code{epi_model} object representing a deterministic \strong{SIR}
(Susceptible–Infectious–Recovered) compartmental epidemic model.

The model describes the spread of an infection in a closed population where
susceptible individuals become infectious and subsequently recover with
permanent immunity.
}
\details{
\subsection{State variables}{

The model is defined in terms of the following state variables:
\describe{
\item{S(t)}{Number of susceptible individuals at time \eqn{t}.}
\item{I(t)}{Number of infectious (actively infected) individuals at time \eqn{t}.}
\item{R(t)}{Number of removed/recovered individuals at time \eqn{t}.}
}

The total population size is conserved:
\deqn{N = S(t) + I(t) + R(t).}
}

\subsection{Model variables}{

The SIR model declares the following variables:
\describe{
\item{\code{"S"}}{Susceptible population size.}
\item{\code{"I"}}{Infectious population size.}
\item{\code{"R"}}{Recovered (removed) population size.}
\item{\code{"incidence"}}{Instantaneous rate of new infections
\eqn{\lambda(t)} returned by the model's right-hand side.}
}

All declared variables may be used as observables in generic utilities
and summary methods built around \code{epi_model} objects.
}

\subsection{Parameters}{

The SIR model depends on the following parameters:
\describe{
\item{beta}{Transmission rate (per day).}
\item{gamma}{Recovery/removal rate (per day).}
}
}

\subsection{Model equations}{

New infections occur at rate
\deqn{\lambda(t) = \beta \frac{S(t)\, I(t)}{N}.}

The system of ordinary differential equations is:
\deqn{
\begin{aligned}
\frac{dS}{dt} &= -\lambda(t), \\
\frac{dI}{dt} &= \lambda(t) - \gamma I(t), \\
\frac{dR}{dt} &= \gamma I(t). \\
\end{aligned}
}
}

\subsection{Usage}{

This predefined model object is intended to be used with generic utilities
such as \code{\link{simulate_epi}}, \code{\link{plot.sim_epi}}, and
\code{\link{summary.sim_epi}} that operate on \code{epi_model} objects.
}
}
\examples{
## Simulate a SIR epidemic
sim <- simulate_epi(
  model = SIR_MODEL,
  times = 0:200,
  parms = c(beta = 0.3, gamma = 0.1),
  init  = c(S = 1e6, I = 10, R = 0)
)

plot(sim)

## Plot incidence
plot(sim, what = "incidence")


}
\seealso{
\code{\link{simulate_epi}},
\code{\link{epi_model}}
}
\keyword{datasets}
