% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODEL_SIRD.R
\docType{data}
\name{SIRD_MODEL}
\alias{SIRD_MODEL}
\title{SIRD epidemic model}
\format{
An object of class \code{"epi_model"}.
}
\usage{
SIRD_MODEL
}
\value{
An object of class \code{"epi_model"} for the predefined SIRD model.
}
\description{
An \code{epi_model} object representing a deterministic \strong{SIRD}
(Susceptible-Infectious-Recovered-Deceased) compartmental epidemic model.
}
\details{
The model is defined by the system:
\deqn{
\begin{aligned}
\frac{dS}{dt} &= -\lambda(t), \\
\frac{dI}{dt} &= \lambda(t) - \gamma I(t) - \mu I(t), \\
\frac{dR}{dt} &= \gamma I(t), \\
\frac{dD}{dt} &= \mu I(t),
\end{aligned}
}
where \eqn{\lambda(t) = \beta\, S(t)\, I(t)/N} and
\eqn{N = S(t) + I(t) + R(t)}.
}
\examples{
sim <- simulate_epi(
  model = SIRD_MODEL,
  times = 0:200,
  parms = c(beta = 0.25, gamma = 0.1, mu = 0.01),
  init = c(S = 1e6, I = 10, R = 0, D = 0)
)

plot(sim)
plot(sim, what = "incidence")

}
\references{
Richard M. (2024). \emph{SIR and variants notes}.
\url{https://www.math.nagoya-u.ac.jp/~richard/teaching/s2024/SML_Vic_2.pdf}
}
\seealso{
\code{\link{simulate_epi}},
\code{\link{epi_model}}
}
\keyword{datasets}
